pyproject.toml
[project]
name = "regimesense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-trade-api>=3.2.0",
    "hmmlearn>=0.3.3",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "yfinance>=0.2.58",
]
